Move `transform_cycles` to `transform` module
